requirements.txt
psutil
pyyaml
llama-cpp-python
faster-whisper
pyttsx3
pytest
